backend/requirements.txt
# Web Framework
flask==3.1.0
flask-cors==5.0.0

# Strands Agents
strands-agents==1.14.0

# MCP and integrations
mcp==1.20.0

# Data validation
pydantic==2.12.3

# HTTP client
httpx==0.28.1

# Environment management
python-dotenv==1.2.1

# ASGI server (optional, for production)
gunicorn==23.0.0
